backend/requirements.txt
Django>=5.2.6
djangorestframework>=3.16.1
django-cors-headers>=4.8.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-factoryboy>=2.6.0
factory-boy>=3.3.0